deduplicate screen_init and display_set_gui_size
