+++ ENIGMAsystem/SHELL/Graphics_Systems/Direct3D11/DX11draw.cpp
@@ -52,7 +52,7 @@ void draw_set_msaa_enabled(bool enable)
 
 }
 
-void draw_enable_alphablend(bool enable) {
+void draw_set_blend(bool enable) {
 
 }
